Create PowerSet(RecursiveApproach).cpp
In the powerset function, we pass the input vector. Inside, we create two vectors: res (a 2D vector) to store the results and current to build each subset. We first add the current subset to the results. As we explore, we include elements in current, and when we reach a certain depth, we push this subset to res. After that, we remove the last element from current (backtracking) to try different combinations. This way, we explore all possible subsets of the input vector.